+++ android/app/src/main/java/com/oppzippy/openscq30/lib/OkPacket.java
@@ -9,7 +9,7 @@ public final class OkPacket {
     }
     private static native long init() throws Exception;
 
-    public static @NonNull java.util.Optional<OkPacket> fromBytes(@NonNull short [] bytes) throws Exception {
+    public static @NonNull java.util.Optional<OkPacket> fromBytes(@NonNull byte [] bytes) throws Exception {
         long ret = do_fromBytes(bytes);
         java.util.Optional<OkPacket> convRet;
         if (ret != 0) {
@@ -20,7 +20,7 @@ public final class OkPacket {
 
         return convRet;
     }
-    private static native long do_fromBytes(short [] bytes) throws Exception;
+    private static native long do_fromBytes(byte [] bytes) throws Exception;
 
     public synchronized void delete() {
         if (mNativeObj != 0) {
