+++ android/app/src/main/java/com/oppzippy/openscq30/lib/SetAmbientSoundModePacket.java
@@ -11,12 +11,12 @@ public final class SetAmbientSoundModePacket {
     }
     private static native long init(int ambient_sound_mode, int noise_canceling_mode);
 
-    public final short [] bytes() {
-        short [] ret = do_bytes(mNativeObj);
+    public final byte [] bytes() {
+        byte [] ret = do_bytes(mNativeObj);
 
         return ret;
     }
-    private static native short [] do_bytes(long self);
+    private static native byte [] do_bytes(long self);
 
     public synchronized void delete() {
         if (mNativeObj != 0) {
