+++ android/app/src/main/java/com/oppzippy/openscq30/lib/SetEqualizerPacket.java
@@ -13,12 +13,12 @@ public final class SetEqualizerPacket {
     }
     private static native long init(long configuration);
 
-    public final short [] bytes() {
-        short [] ret = do_bytes(mNativeObj);
+    public final byte [] bytes() {
+        byte [] ret = do_bytes(mNativeObj);
 
         return ret;
     }
-    private static native short [] do_bytes(long self);
+    private static native byte [] do_bytes(long self);
 
     public synchronized void delete() {
         if (mNativeObj != 0) {
